Refactor authentication dependencies
Move the authentication dependencies under gafaelfawr.dependencies
and out of the module that has the HTTP authentication types and
infrastructure.  This makes the module layout a bit more consistent.

Also update comments to remove old aiohttp references.
